fix(table-views): key demotion off URL viewId only; move preset catalog to domain layer (LFE-14699)
The sessionStorage storedViewId is deliberately not synced when a link
with explicit table state is opened, so it can go stale (chip applied
earlier in the tab, then a saved view's permalink opened). Keying the
demotion off it would strip the user view's ?viewId on the next edit —
read the URL's selectedViewId only, which every intended demotion path
sets. Adds a stale-storage regression test.

Move the system preset catalog into the domain layer (server service
re-exports it unchanged) so the round-trip clienttest reads it from the
client-safe root barrel instead of the server barrel.
