+++ packages/shared/src/domain/index.ts
@@ -3,6 +3,7 @@ export * from "./observations";
 export * from "./traces";
 export * from "./scores";
 export * from "./table-view-presets";
+export * from "./table-view-presets-catalog";
 export * from "./automations";
 export * from "./webhooks";
 export * from "./prompts";
